requirements.txt
# Core dependencies
fastmcp>=0.1.0
httpx>=0.24.0
pydantic>=2.0.0

# Notifications
python-telegram-bot>=20.0
discord.py>=2.0.0

# Infrastructure APIs
proxmoxer>=2.0.0
# routeros-api (for MikroTik)

# Database
# sqlite3 (built-in)

